requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
asttokens==2.2.1
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
distconfig3==1.0.1
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
packaging==23.1
pluggy==1.2.0
pycryptodome==3.18.0
PyHamcrest==2.0.4
pytest==7.4.0
PyYAML==6.0.1
requests==2.31.0
semantic-version==2.10.0
six==1.16.0
toml==0.10.2
urllib3==2.0.4
vyper-config==1.0.0
watchdog==3.0.0
zipp==3.16.2
